+++ sample/src/Launcher.kt
@@ -60,7 +60,7 @@ fun sampleServer(port: Int = 8080) = Server(listen = InetSocketAddress(port)).ap
 
   context("/api") {
     useOnly<JsonBody>() // in case only json should be supported in this context
+    useHashCodeAsETag() // automatically send 304 NotModified if request generates the same response as before
     annotated<MyRoutes>() // read routes from an annotated class - such classes are easier to unit-test
   }
 }
-